refactor: refactor MyFrom into two components
