+++ src/App.js
@@ -1,6 +1,6 @@
 import "./App.css";
 import { useState } from "react";
-import MyForm from "./components/MyForm";
+import MyFormWithFormik from "./components/MyForm/MyFromWithFormik";
 import MyTable from "./components/MyTable";
 
 function App() {
@@ -13,7 +13,7 @@ function App() {
     <div
       style={{ backgroundColor: "rgba(118, 118, 118, 0.5)", height: "100vh" }}
     >
-      <MyForm onSubmit={saveData} />
+      <MyFormWithFormik onSubmit={saveData} />
       <MyTable data={data} />
     </div>
   );
